requirements.txt
APScheduler==3.8.0
Flask==2.0.1
Pillow==8.3.1
pyjokes==0.6.0
python-dotenv==0.19.0
requests==2.22.0
requests-oauthlib==1.3.0
tweepy==3.9.0